requirements.txt
openai==0.27.8
biopython==1.81

